Refactor smiles drawer options
Add two sub options molecule / reaction
Refactor settings for better extensibility
